L14: enable -Wconversion/-Wsign-conversion and fix narrowing findings
Add -Wconversion -Wsign-conversion to the shared drivers.cmake warning
set and every project's target block. Fix the legitimate narrowing /
double-to-float findings so all five ARM Debug targets build clean:
explicit static_casts in LIS3DSH/Dac/I2C/Rtc/BasicTimer/PWM/GenericTimer
and the heap/stack utilities, uint16_t length in SPI_arbiter, (char) on
_read, int16_t motion-sample assembly, and f-suffixed float literals in
the timer Configs and the AppLogic scale constant. No API or behaviour
change; native root suite green (411/411).

+++ drivers/components/LIS3DSH/LIS3DSH.cpp
@@ -374,7 +374,7 @@ bool LIS3DSH::Configure(const Config& cfg)
 
     const uint8_t BDU = (mUseHardwareFifo) ? 0 : 1;     // 0: disabled (default if fifo is used), 1: enabled
 
-    uint8_t src = (ODR | (BDU << 3) | AXES_ENABLED);    // Set sample frequency, all axes enabled, not using BDU
+    uint8_t src = static_cast<uint8_t>(ODR | (BDU << 3) | AXES_ENABLED);    // Set sample frequency, all axes enabled, not using BDU
     bool result = WriteRegister(CTRL_REG4, &src, 1);
     EXPECT(result);
