+++ drivers/drivers/PWM/PWM.cpp
@@ -263,7 +263,7 @@ uint16_t PWM::CalculatePeriod(float desiredFrequency)
 
     // timer_period = (timer_tick_frequency / PWM_frequency) - 1
 
-    uint32_t timer_period = (tick / desiredFrequency) - 1;
+    uint32_t timer_period = static_cast<uint32_t>(static_cast<float>(tick) / desiredFrequency - 1.0f);
 
     // Cap at UINT16_MAX. TIM2/5 ARR is 32-bit but TIM3/4 is 16-bit; use the
     // common cap so the same Config range applies on every supported instance.
@@ -295,7 +295,7 @@ uint32_t PWM::CalculatePulse(float desiredDutyCycle, uint32_t period)
     // CNT never reaches CCR, output stays inactive, and with the inverted
     // OCPolarity below the channel sticks at the user's "ON" level -- i.e.
     // 0% silently turns into 100%. Special-case it.
-    const uint32_t scaled = static_cast<uint32_t>((period + 1U) * desiredDutyCycle);
+    const uint32_t scaled = static_cast<uint32_t>(static_cast<float>(period + 1U) * desiredDutyCycle);
     if (scaled == 0U) { return 0U; }
 
     // pulse_length = (timer_period + 1) * duty_cycle - 1
